grantai-pro/requirements.txt
Flask==2.3.3
pymongo==4.5.0
stripe==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
requests==2.31.0
dnspython==2.4.2
pymongo[srv]==4.5.0
